requirements.txt
streamlit
pdfplumber
pytesseract
pdf2image
google-generativeai
Pillow
python-dotenv